requirements.txt
ipykernel==6.21.*
psycopg2-binary==2.9.*
pandas==1.5.*
matplotlib==3.7.*
seaborn==0.12.*
statsmodels==0.13.*
beautifulsoup4==4.12.*
gensim==4.3.*
faiss-cpu==1.7.*
adjustText==0.8.*
scikit-learn==1.2.*